Changed/removed a whole bunch of obsolete code and config
