--- CHANGES.rst
@@ -1,4 +1,0 @@
-As of 1/18/17:
-- Changed ES mapping and embedding system to resolve field conflicts with ES versions 2 and up.
-- Selective embedding of objects in snovault based of the /types/ file of the object.
-- Tests to accompany these changes.
