--- buildout-travis.cfg
@@ -1,4 +1,0 @@
-[buildout]
-extends = buildout.cfg
-parts -=
-    aws
