+++ buildout.cfg
@@ -34,7 +34,6 @@ zope.sqlalchemy = git https://github.com/zopefoundation/zope.sqlalchemy.git
 [versions]
 # Hand set versions
 pyramid = 1.10.1
-# Update .travis.yml and cloud-config.yml when updating buildout
 six = 1.11.0
 
 [snovault]
