--- cloud-config-elasticsearch.yml
@@ -1,152 +1,0 @@
-#cloud-config
-
-# Instance
-
-apt_sources:
-- source: "ppa:webupd8team/java"
-- source: "deb http://packages.elasticsearch.org/elasticsearch/1.7/debian stable main"
-  key: |
-    -----BEGIN PGP PUBLIC KEY BLOCK-----
-    Version: GnuPG v2.0.14 (GNU/Linux)
-
-    mQENBFI3HsoBCADXDtbNJnxbPqB1vDNtCsqhe49vFYsZN9IOZsZXgp7aHjh6CJBD
-    A+bGFOwyhbd7at35jQjWAw1O3cfYsKAmFy+Ar3LHCMkV3oZspJACTIgCrwnkic/9
-    CUliQe324qvObU2QRtP4Fl0zWcfb/S8UYzWXWIFuJqMvE9MaRY1bwUBvzoqavLGZ
-    j3SF1SPO+TB5QrHkrQHBsmX+Jda6d4Ylt8/t6CvMwgQNlrlzIO9WT+YN6zS+sqHd
-    1YK/aY5qhoLNhp9G/HxhcSVCkLq8SStj1ZZ1S9juBPoXV1ZWNbxFNGwOh/NYGldD
-    2kmBf3YgCqeLzHahsAEpvAm8TBa7Q9W21C8vABEBAAG0RUVsYXN0aWNzZWFyY2gg
-    KEVsYXN0aWNzZWFyY2ggU2lnbmluZyBLZXkpIDxkZXZfb3BzQGVsYXN0aWNzZWFy
-    Y2gub3JnPokBOAQTAQIAIgUCUjceygIbAwYLCQgHAwIGFQgCCQoLBBYCAwECHgEC
-    F4AACgkQ0n1mbNiOQrRzjAgAlTUQ1mgo3nK6BGXbj4XAJvuZDG0HILiUt+pPnz75
-    nsf0NWhqR4yGFlmpuctgCmTD+HzYtV9fp9qW/bwVuJCNtKXk3sdzYABY+Yl0Cez/
-    7C2GuGCOlbn0luCNT9BxJnh4mC9h/cKI3y5jvZ7wavwe41teqG14V+EoFSn3NPKm
-    TxcDTFrV7SmVPxCBcQze00cJhprKxkuZMPPVqpBS+JfDQtzUQD/LSFfhHj9eD+Xe
-    8d7sw+XvxB2aN4gnTlRzjL1nTRp0h2/IOGkqYfIG9rWmSLNlxhB2t+c0RsjdGM4/
-    eRlPWylFbVMc5pmDpItrkWSnzBfkmXL3vO2X3WvwmSFiQbkBDQRSNx7KAQgA5JUl
-    zcMW5/cuyZR8alSacKqhSbvoSqqbzHKcUQZmlzNMKGTABFG1yRx9r+wa/fvqP6OT
-    RzRDvVS/cycws8YX7Ddum7x8uI95b9ye1/Xy5noPEm8cD+hplnpU+PBQZJ5XJ2I+
-    1l9Nixx47wPGXeClLqcdn0ayd+v+Rwf3/XUJrvccG2YZUiQ4jWZkoxsA07xx7Bj+
-    Lt8/FKG7sHRFvePFU0ZS6JFx9GJqjSBbHRRkam+4emW3uWgVfZxuwcUCn1ayNgRt
-    KiFv9jQrg2TIWEvzYx9tywTCxc+FFMWAlbCzi+m4WD+QUWWfDQ009U/WM0ks0Kww
-    EwSk/UDuToxGnKU2dQARAQABiQEfBBgBAgAJBQJSNx7KAhsMAAoJENJ9ZmzYjkK0
-    c3MIAIE9hAR20mqJWLcsxLtrRs6uNF1VrpB+4n/55QU7oxA1iVBO6IFu4qgsF12J
-    TavnJ5MLaETlggXY+zDef9syTPXoQctpzcaNVDmedwo1SiL03uMoblOvWpMR/Y0j
-    6rm7IgrMWUDXDPvoPGjMl2q1iTeyHkMZEyUJ8SKsaHh4jV9wp9KmC8C+9CwMukL7
-    vM5w8cgvJoAwsp3Fn59AxWthN3XJYcnMfStkIuWgR7U2r+a210W6vnUxU4oN0PmM
-    cursYPyeV0NX/KQeUeNMwGTFB6QHS/anRaGQewijkrYYoTNtfllxIu9XYmiBERQ/
-    qPDlGRlOgVTd9xUfHFkzB52c70E=
-    =92oX
-    -----END PGP PUBLIC KEY BLOCK-----
-
-bootcmd:
-- set -ex
-- cloud-init-per once ssh-users-ca echo "TrustedUserCAKeys /etc/ssh/users_ca.pub" >> /etc/ssh/sshd_config
-- cloud-init-per once accepted-oracle-license-v1-1 echo "oracle-java8-installer shared/accepted-oracle-license-v1-1 select true" | debconf-set-selections
-- cloud-init-per once fallocate-swapfile fallocate -l 4G /swapfile
-- cloud-init-per once chmod-swapfile chmod 600 /swapfile
-- cloud-init-per once mkswap-swapfile mkswap /swapfile
-- MEMGIGS=$(awk '/MemTotal/{printf int($2 / 1024**2)}' /proc/meminfo)
-- if [ "$MEMGIGS" -gt 12 ];
-- then
--    echo "ES_HEAP_SIZE=8g" > /etc/default/elasticsearch
-- else
--    echo "ES_HEAP_SIZE=4g" > /etc/default/elasticsearch
--    sysctl "vm.swappiness=1"
--    swapon /swapfile
-- fi
-
-package_upgrade: true
-
-packages:
-- build-essential
-- elasticsearch
-- libssl-dev
-- oracle-java8-installer
-- oracle-java8-set-default
-- unattended-upgrades
-
-power_state:
-  mode: reboot
-
-output:
-  all: '| tee -a /var/log/cloud-init-output.log'
-
-runcmd:
-# Ideally this would build as a different user so snowflakes only has read
-# permissions
-- set -ex
-- update-rc.d elasticsearch defaults
-- sudo bash /etc/elasticsearch/cluster.sh %(CLUSTER_NAME)s
-- sudo /usr/share/elasticsearch/bin/plugin install elasticsearch/elasticsearch-cloud-aws/2.7.1
-- service elasticsearch start
-- mkdir /opt/cloudwatchmon
-- chown build:build /opt/cloudwatchmon
-- sudo -u build virtualenv --python=python2.7 /opt/cloudwatchmon
-- sudo -u build /opt/cloudwatchmon/bin/pip install -r cloudwatchmon-requirements.txt
-
-
-users:
-- default
-- name: build
-  gecos: Build user
-  inactive: true
-  system: true
-- name: snowflakes
-  gecos: Snowflake Database daemon user
-  inactive: true
-  system: true
-  # Specified homedir must exist
-  # https://github.com/rubygems/rubygems/issues/689
-  homedir: /srv/snowflakes
-
-write_files:
-- path: /etc/apt/apt.conf.d/20auto-upgrades
-  content: |
-    APT::Periodic::Update-Package-Lists "1";
-    APT::Periodic::Unattended-Upgrade "1";
-
-- path: /etc/apt/apt.conf.d/50unattended-upgrades
-  content: |
-    Unattended-Upgrade::Allowed-Origins {
-        "${distro_id} ${distro_codename}-security";
-    };
-    Unattended-Upgrade::Automatic-Reboot "true";
-
-- path: /etc/cron.d/cloudwatchmon
-  content: |
-    */5 * * * * nobody /opt/cloudwatchmon/bin/mon-put-instance-stats.py --mem-util --swap-util --disk-space-util --disk-path=/ --from-cron
-
-- path: /etc/elasticsearch/elasticsearch.yml
-  content: |
-    index.search.slowlog.threshold.query.warn: 8s
-    index.search.slowlog.threshold.fetch.warn: 1s
-    index.indexing.slowlog.threshold.index.warn: 25s
-    network.host: 0.0.0.0
-    action.auto_create_index: false
-    index.mapper.dynamic: false
-
-    node.master: false
-    node.data: true
-
-    discovery:
-        type: ec2
-
-    cloud.aws.region: us-west-2
-
-    discovery.ec2.groups: elasticsearch-https, ssh-http-https
-
-    # discovery.zen.minimum_master_nodes: 2
-
-- path: /etc/elasticsearch/cluster.sh
-  content: |
-    #!/bin/bash
-    name=$1
-    
-    if [[ -n "$name" ]]; then
-        echo "cluster.name: $name" >> /etc/elasticsearch/elasticsearch.yml
-    else
-        echo "argument error"
-    fi
-
-- path: /etc/ssh/users_ca.pub
-  content: ssh-rsa AAAAB3NzaC1yc2EAAAABIwAAAQEAv/ymOcnN4LhM4NACc3Or116XXJ6KytuOgB/+1qNkOFBqBosrn7cmJ35rsoNHRgYNrCsRE9ch74RKsN6H72FtSJgBhGh/9oUK7Os6Fqt3/ZZXxgxIx6ubs/MTgrxrAnujiBxUXMXQhLKMriNMpo8mt4nGYVtLk9PBjiyfncaS8H9ZKoNio9dhP8bmTuYvioAI35dqKdSlVLyzr/XkZxia8Ki+pQ0N6uuiEwMR3ToM+LSp8wpFOOAiu4PEAujRW7us/+1hlpKWfn0J7/V3826joHE+I967Vg/+ikcVhF77JjK1nib879VgCWfmn1HPQosIpk4yJfVgGvRVI7I2nfBPVw== snowflakes@snowflakes.org
