--- demo.cfg
@@ -1,2 +1,0 @@
-[buildout]
-extends = buildout.cfg
