+++ docs/source/index.rst
@@ -1,18 +1,18 @@
 ========================
-SnoVault
+Snovault
 ========================
 
-Snovault is a JSON-LD Database Framework that serves as the backend for the 4DN Data portal and CGAP. Version 0.2
+Snovault is a JSON-LD Database Framework that serves as the backend for the 4DN Data portal and CGAP.
 
 |Build status|_
 
-.. |Build status| image:: https://travis-ci.org/ENCODE-DCC/snovault.png?branch=master
-.. _Build status: https://travis-ci.org/ENCODE-DCC/snovault
+.. |Build status| image:: https://travis-ci.org/4dn-dcic/snovault.svg?branch=master
+.. _Build status: https://travis-ci.org/4dn-dcic/snovault
 
 Installation Instructions
 =========================
 
-Currently these are for Mac OSX.  For linux, look at cloud-config.yml it should be reasonably easy to infer from that
+Currently these are for Mac OSX using homebrew. If using linux, install dependencies with a different package manager.
 
 Step 0: Install Xcode (from App Store) and homebrew: http://brew.sh::
 
@@ -23,10 +23,10 @@ Step 1: Verify that homebrew is working properly::
 
 Step 2: Install or update dependencies::
 
-    $ brew install libevent libmagic libxml2 libxslt openssl postgresql graphviz nginx python3
+    $ brew install libevent libmagic libxml2 libxslt openssl postgresql graphviz python3
     $ brew install freetype libjpeg libtiff littlecms webp  # Required by Pillow
     $ brew tap homebrew/versions
-    $ brew install elasticsearch@5.6 node@10
+    $ brew install elasticsearch@5.6
 
 If you need to update dependencies::
 
@@ -54,25 +54,6 @@ If you wish to completely rebuild the application, or have updated dependencies:
 
     Then goto Step 3.
 
-Step 4: Start the application locally
-
-In one terminal startup the database servers and nginx proxy with::
-
-    $ bin/dev-servers development.ini --app-name app --clear --init --load
-
-This will first clear any existing data in /tmp/snowflakes
-Then postgres and elasticsearch servers will be initiated within /tmp/snowflakes.
-An nginx proxy running on port 8000 will be started.
-The servers are started, and finally the test set will be loaded.
-
-In a second terminal, run the app with::
-
-    $ bin/pserve development.ini
-
-Indexing will then proceed in a background thread similar to the production setup.
-
-Browse to the interface at http://localhost:8000/.
-
 
 Running tests
 =============
@@ -92,47 +73,3 @@ Specific tests to run locally for schema changes::
 Run the Pyramid tests with::
 
     $ bin/test
-
-Run the Javascript tests with::
-
-    $ npm test
-
-Or if you need to supply command line arguments::
-
-    $ ./node_modules/.bin/jest
-
-
-SublimeLinter
-=============
-
-To setup SublimeLinter with Sublime Text 3, first install the linters::
-
-    $ easy_install-2.7 flake8
-    $ npm install -g jshint
-    $ npm install -g jsxhint
-
-After first setting up `Package Control`_ (follow install and usage instructions on site), use it to install the following packages in Sublime Text 3:
-
-    * sublimelinter
-    * sublimelinter-flake8
-    * sublimelinter-jsxhint
-    * jsx
-    * sublimelinter-jshint
-
-.. _`Package Control`: https://sublime.wbond.net/
-
-*Development Contents:*
-
- .. toctree::
-   :maxdepth: 4
-
-   self
-   overview
-   testing
-   auth
-   custom-travis
-   embedding-and-indexing
-   invalidation
-   object-lifecycle
-   search_info
-   snowflakes
