+++ src/snovault/tests/conftest.py
@@ -8,6 +8,23 @@ pytest_plugins = [
 ]
 
 
+def pytest_addoption(parser):
+    parser.addoption("--es", action="store", default="", dest='es',
+        help="use a remote es for testing")
+    parser.addoption("--aws-auth", action="store_true",
+        help="connect using aws authorization")
+
+
+@pytest.fixture(scope='session')
+def remote_es(request):
+    return request.config.getoption("--es")
+
+
+@pytest.fixture(scope='session')
+def aws_auth(request):
+    return request.config.getoption("--aws-auth")
+
+
 # required so that db transactions are properly rolled back in tests
 @pytest.fixture(autouse=True)
 def autouse_external_tx(external_tx):
