--- wal-e-requirements.txt
@@ -1,44 +1,0 @@
-Babel==2.2.0
-argparse==1.2.1
-azure==1.0.3
-azure-common==1.0.0
-azure-mgmt==0.20.2
-azure-mgmt-common==0.20.0
-azure-mgmt-compute==0.20.1
-azure-mgmt-network==0.20.1
-azure-mgmt-nspkg==1.0.0
-azure-mgmt-resource==0.20.1
-azure-mgmt-storage==0.20.0
-azure-nspkg==1.0.0
-azure-servicebus==0.20.1
-azure-servicemanagement-legacy==0.20.2
-azure-storage==0.20.3
-debtcollector==1.3.0
-envfile==1.0
-funcsigs==0.4
-futures==3.0.5
-gevent==1.0.2
-greenlet==0.4.9
-iso8601==0.1.11
-keystoneauth1==2.3.0
-monotonic==0.6
-msgpack-python==0.4.7
-netaddr==0.7.18
-netifaces==0.10.4
-oslo.config==3.9.0
-oslo.i18n==3.4.0
-oslo.serialization==2.4.0
-oslo.utils==3.7.0
-pbr==1.8.1
-positional==1.0.1
-prettytable==0.7.2
-python-dateutil==2.4.2
-python-keystoneclient==2.3.0
-python-swiftclient==2.7.0
-pytz==2015.7
-requests==2.9.1
-six==1.10.0
-stevedore==1.12.0
-wal-e==0.8.1
-wrapt==1.10.6
-wsgiref==0.1.2
